requirements.txt
selenium==4.15.0
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1
openai==1.3.7
python-dotenv==1.0.0
streamlit==1.28.1
plotly==5.17.0